+++ feature/summary/src/main/java/com/githukudenis/summary/navigation/SummaryNavigation.kt
@@ -2,17 +2,26 @@ package com.githukudenis.summary.navigation
 
 import android.os.Build
 import androidx.annotation.RequiresApi
+import androidx.compose.material3.SnackbarHostState
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.githukudenis.summary.ui.SummaryRoute
+import com.githukudenis.summary.ui.home.SummaryRoute
 
 const val summaryNavigationRoute = "summary"
 
 @RequiresApi(Build.VERSION_CODES.O)
 fun NavGraphBuilder.summaryScreen(
-    onOpenHabitDetails: (Long) -> Unit
+    snackbarHostState: SnackbarHostState,
+    onOpenHabitDetails: (Long) -> Unit,
+    onNavigateUp: () -> Unit,
+    onOpenActivity: () -> Unit
 ) {
     composable(route = summaryNavigationRoute) {
-        SummaryRoute(onOpenHabitDetails = onOpenHabitDetails)
+        SummaryRoute(
+            snackbarHostState = snackbarHostState,
+            onOpenHabitDetails = onOpenHabitDetails,
+            onNavigateUp = onNavigateUp,
+            onOpenActivity = onOpenActivity
+        )
     }
 }
