--- feature/summary/src/main/java/com/githukudenis/summary/ui/SummaryUiEvent.kt
@@ -1,7 +1,0 @@
-package com.githukudenis.summary.ui
-
-sealed interface SummaryUiEvent {
-    object Refresh : SummaryUiEvent
-    data class ShowError(val error: UserError) : SummaryUiEvent
-    data class CheckHabit(val habitId: Long) : SummaryUiEvent
-}
