--- feature/summary/src/main/java/com/githukudenis/summary/ui/UserError.kt
@@ -1,12 +1,0 @@
-package com.githukudenis.summary.ui
-
-data class UserError(
-    val id: Long = 0,
-    val message: String? = null,
-    val errorType: ErrorType = ErrorType.DEFAULT
-)
-
-enum class ErrorType {
-    DEFAULT,
-    CRITICAL
-}
