+++ feature/summary/src/main/java/com/githukudenis/summary/ui/components/PersonalizeSheet.kt
@@ -1,0 +1,21 @@
+package com.githukudenis.summary.ui.components
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.runtime.Composable
+import kotlinx.coroutines.launch
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun PersonalizeSheet(
+    onDismiss: () -> Unit,
+    content: @Composable () -> Unit
+) {
+    val bottomSheetState = rememberModalBottomSheetState()
+    ModalBottomSheet(
+        sheetState = bottomSheetState,
+        onDismissRequest = { onDismiss() }) {
+        content()
+    }
+}
