+++ feature/summary/src/main/java/com/githukudenis/summary/ui/home/SummaryUiEvent.kt
@@ -1,0 +1,12 @@
+package com.githukudenis.summary.ui.home
+
+import com.githukudenis.summary.ui.UserMessage
+
+sealed interface SummaryUiEvent {
+    data object Refresh : SummaryUiEvent
+    data class ShowMessage(val error: UserMessage) : SummaryUiEvent
+    data class CheckHabit(val habitId: Long) : SummaryUiEvent
+    data class DismissMessage(val messageId: Long): SummaryUiEvent
+
+    data class EditHabit(val habitId: Long): SummaryUiEvent
+}
